SCRUM-68

+++ WalletService.Models/DTO/WalletModel1BDto/BalanceInformationModel1BDto.cs
@@ -8,6 +8,6 @@ public class BalanceInformationModel1BDto
 
     public decimal? AvailableBalance { get; set; }
 
-    public double? TotalCommissionsPaid { get; set; }
+    public decimal? TotalCommissionsPaid { get; set; }
     public decimal? ServiceBalance { get; set; }
 }
